Clean up JSON output by making self.assets the only public dict in Context. There were other unnecessary assets lists, which greatly cluttered the JSON output with redundant info.
Note that this also requires a new version of the extraction framework to handle the case where self.assets is a dict and not a list.

The _referenced_chunks (formerly referenced_chunks) dict is still required, becuase not all asset headers declared in the file (and thus stored in the self.assets dict) will actually have their asset data in this file - the data could be contained in INSTALL.CXT instead. But since the _referenced_chunks is now "protected" (using C terminology), it doesn't show up in the JSON output.