Rename internal variables to prevent JSON export.
I just use a single underscore to indicate loosely "protected" variables (those that shouldn't be included in a JSON export). Maybe this isn't the best approach because it mixes the concepts of data visibility in classes with data included in the JSON export, but it was definitely the easiest at the time since the jsons library has a way to ignore private/protected class members but not another easy way to ignore arbitrary class members.